refactor tests for new component
