+++ web/html/src/components/input/check/Check.test.tsx
@@ -1,31 +1,12 @@
-import { click, render, screen } from "utils/test-utils";
+import { render, screen } from "utils/test-utils";
 
 import { Check } from "./Check";
 
 describe("Check", () => {
-  test("syncs the indeterminate DOM property to true", () => {
-    render(<Check indeterminate readOnly />);
+  test("applies the default input class to the input", () => {
+    render(<Check readOnly />);
 
-    expect((screen.getByRole("checkbox") as HTMLInputElement).indeterminate).toBe(true);
-  });
-
-  test("clears the indeterminate DOM property when the prop becomes undefined", () => {
-    const { rerender } = render(<Check indeterminate readOnly />);
-    expect((screen.getByRole("checkbox") as HTMLInputElement).indeterminate).toBe(true);
-
-    rerender(<Check readOnly />);
-
-    expect((screen.getByRole("checkbox") as HTMLInputElement).indeterminate).toBe(false);
-  });
-
-  test("invokes onChange with the new boolean value when toggled", async () => {
-    const onChange = jest.fn();
-    render(<Check checked={false} onChange={onChange} />);
-
-    await click(screen.getByRole("checkbox"));
-
-    expect(onChange).toHaveBeenCalledTimes(1);
-    expect(onChange).toHaveBeenCalledWith(true);
+    expect(screen.getByRole("checkbox").className).toContain("form-check-input");
   });
 
   test("applies a custom inputClass to the input, overriding the default", () => {
@@ -40,5 +21,11 @@ describe("Check", () => {
     const { container } = render(<Check readOnly />);
 
     expect(container.querySelector("label")).not.toBeNull();
+  });
+
+  test("applies className to the label wrapper", () => {
+    const { container } = render(<Check className="wrapper-check" readOnly />);
+
+    expect(container.querySelector("label")?.className).toContain("wrapper-check");
   });
 });
